+++ CHANGELOG.md
@@ -181,10 +181,15 @@
 * Create Jest/RTL test for `ChargeForm`. Refs UIU-2260.
 * Create Jest/RTL test for `BulkOverrideLoansList`. Refs UIU-2328.
 * Create Jest/RTL test for `BulkOverrideInfo`. Refs UIU-2320.
+* Create Jest/RTL test for `Menu`. Refs UIU-2357.
 * Create Jest/RTL test for `memoize`. Refs UIU-2338.
 * Create Jest/RTL test for `withServicePoints`. Refs UIU-2279.
 * Create Jest/RTL test for `ErrorPane`. Refs UIU-2339.
 * Create Jest/RTL test for `BulkOverrideInfo`. Refs UIU-2320.
+* Create Jest/RTL test for `ConditionsForm`. Refs UIU-2382.
+* Create Jest/RTL test for `Conditions`. Refs UIU-2405.
+* Create Jest/RTL test for `getListPresentation`. Refs UIU-2298.
+* Create Jest/RTL test for `getListDataFormatter`. Refs UIU-2351.
 
 ## [6.1.0](https://github.com/folio-org/ui-users/tree/v6.1.0) (2021-06-18)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v6.0.0...v6.1.0)